+++ modules/sdk-coin-ton/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.1](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-ton@3.2.0...@bitgo/sdk-coin-ton@3.2.1) (2024-07-30)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-ton
+
 # [3.2.0](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-ton@3.1.0...@bitgo/sdk-coin-ton@3.2.0) (2024-07-24)
 
 ### Features