+++ modules/sdk-coin-xrp/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.9](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-xrp@2.1.8...@bitgo/sdk-coin-xrp@2.1.9) (2024-07-30)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-xrp
+
 ## [2.1.8](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-xrp@2.1.7...@bitgo/sdk-coin-xrp@2.1.8) (2024-07-24)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-xrp